Fix softlock when automining with bone pickaxe
It seems that the pickaxe use action does not play well with an item having charges, meaning for the item to still exist at all it's going to have to be infinite-use. Defeats the point of it being consumable but neither BN nor DDA has good support for consumable tools, so this is a problem that's hard to avoid without workarounds.